creates two views of the energyBalance graph
