--- app/frontend/components/EnergyBalance/testData.tsx
@@ -1,31 +1,0 @@
-const changedDevelopments = [
-    {
-        sd: 'capacity_of_rooftop_pv_in_households',
-        gid: 'GM0575',
-        inputs: [
-            {
-                key: 'capacity_of_rooftop_pv_in_households_proj1_cv_ketels_methaan',
-                value: 49,
-            },
-            {
-                key: 'capacity_of_rooftop_pv_in_households_proj2_cv_ketels_methaan',
-                value: 55,
-            },
-        ],
-        projId: 'proj1',
-    },
-    {
-        cd: 'capacity_of_rooftop_pv_in_households',
-        gid: 'GM0575',
-        inputs: [
-            {
-                key: 'capacity_of_rooftop_pv_in_households_proj1_cv_ketels_methaan',
-                value: 49,
-            },
-            {
-                key: 'capacity_of_rooftop_pv_in_households_proj2_cv_ketels_methaan',
-                value: 55,
-            },
-        ],
-    },
-];
